Refactor PlexWatch event processing and clean up codebase
This commit majorly refactors the event processing, improving the overall efficiency of the system and aiding in preventing event stacking. The StreamStartedSubscription and TranscodeChangedSubscription classes have been deleted and their functionality is now directed to and handled by the SubscriptionActions class. Minor changes are also made throughout the codebase to simplify and clean up the existing structure.