Rename 'Session' to 'SessionKey' across various classes
The commit renames the 'Session' property to 'SessionKey' in several classes such as StreamStartedEvent, TranscodeChangedEvent, and in the EventParsingService. It also updates any related log information accordingly. Additionally, it introduces a check in the TranscodeChecker utility to detect if the session has 'transcode' in the audio or video decision when the quality is set to 'original'.